Update docker-image.yml
n this modified workflow, the on field is updated to trigger the workflow when a new tag is pushed or a new release is created. The VERSION in the Prepare step is updated to use the tag or release name. This will result in Docker images tagged with the tag or release name, and you can pull the image with the command docker pull adabalap/vulcan:v1.0.0 where v1.0.0 is the tag or release name.